requirements.txt
schwab-py==1.4.0
anthropic==0.40.0
openai>=1.0.0
discord.py==2.4.0
yfinance==0.2.51
pandas==2.2.3
lxml>=5.0.0
tenacity>=8.0.0
APScheduler==3.10.4
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
